integrations/marker-extension/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aspose-slides-marker-extension"
dynamic = ["version"]
dependencies = [
  "marker-pdf",
  "aspose.slides>=25.7.0"
]

[tool.hatch.version]
path = "src/aspose_slides_marker_extension/__about__.py"
